Work in progress: messy fix to work around uintptr_t definition in some tests, change definition from unit to u8 to make pointer arithmetic work sensibly

+++ tests/ui/abi/extern/extern-crosscrate.rs
@@ -9,13 +9,13 @@ extern crate libc;
 
 fn fact(n: libc::uintptr_t) -> libc::uintptr_t {
     unsafe {
-        println!("n = {}", n);
+        println!("n = {}", n as usize);
         externcallback::rustrt::rust_dbg_call(externcallback::cb, n)
     }
 }
 
 pub fn main() {
-    let result = fact(10);
-    println!("result = {}", result);
-    assert_eq!(result, 3628800);
+    let result = fact(10 as libc::uintptr_t);
+    println!("result = {}", result as usize);
+    assert_eq!(result, 3628800 as libc::uintptr_t);
 }
